API実装、テーブル登録

+++ db/migrate/20180118065931_create_similar_artists.rb
@@ -3,7 +3,7 @@ class CreateSimilarArtists < ActiveRecord::Migration
     create_table :similar_artists do |t|
       t.integer :artist_id
       t.string :name
-      t.integer :match
+      t.float :match
 
       t.timestamps
     end
